[I-6] Set Up MongoDB and .gitignore
This commit set up node.js and mongodb as well as created a gitignore.  It is unclear whether I should include package.json in the gitignore.  The community seems to want to include it, but the NPM docs say that it can't be ignored by .npmignore. It is currently being ignored, though, so I will have to monitor any performance issues as a result of this.